Improvement attempts based on feedback fixing things like : file name, equals operator format, date extraction and more.
